Rename wrap-multilines to jsx-wrap-multilines
This rule is JSX-specific, not React-specific. Changing its name makes
this clearer. I've kept the old one around as deprecated. We can remove
it at the next major version bump.

Addresses #686. Fixes #668.
