+++ lib/rules/wrap-multilines.js
@@ -1,89 +1,32 @@
 /**
  * @fileoverview Prevent missing parentheses around multilines JSX
  * @author Yannick Croissant
+ * @deprecated
  */
 'use strict';
-
-// ------------------------------------------------------------------------------
-// Constants
-// ------------------------------------------------------------------------------
-
-var DEFAULTS = {
-  declaration: true,
-  assignment: true,
-  return: true
-};
 
 // ------------------------------------------------------------------------------
 // Rule Definition
 // ------------------------------------------------------------------------------
 
+var util = require('util');
+var jsxWrapMultilines = require('./jsx-wrap-multilines');
+var isWarnedForDeprecation = false;
+
 module.exports = function(context) {
-
-  var sourceCode = context.getSourceCode();
-
-  function isParenthesised(node) {
-    var previousToken = sourceCode.getTokenBefore(node);
-    var nextToken = sourceCode.getTokenAfter(node);
-
-    return previousToken && nextToken &&
-      previousToken.value === '(' && previousToken.range[1] <= node.range[0] &&
-      nextToken.value === ')' && nextToken.range[0] >= node.range[1];
-  }
-
-  function isMultilines(node) {
-    return node.loc.start.line !== node.loc.end.line;
-  }
-
-  function check(node) {
-    if (!node || node.type !== 'JSXElement') {
-      return;
-    }
-
-    if (!isParenthesised(node) && isMultilines(node)) {
-      context.report({
-        node: node,
-        message: 'Missing parentheses around multilines JSX',
-        fix: function(fixer) {
-          return fixer.replaceText(node, '(' + sourceCode.getText(node) + ')');
-        }
-      });
-    }
-  }
-
-  function isEnabled(type) {
-    var userOptions = context.options[0] || {};
-    if (({}).hasOwnProperty.call(userOptions, type)) {
-      return userOptions[type];
-    }
-    return DEFAULTS[type];
-  }
-
-  // --------------------------------------------------------------------------
-  // Public
-  // --------------------------------------------------------------------------
-
-  return {
-
-    VariableDeclarator: function(node) {
-      if (isEnabled('declaration')) {
-        check(node.init);
+  return util._extend(jsxWrapMultilines(context), {
+    Program: function() {
+      if (isWarnedForDeprecation || /\=-(f|-format)=/.test(process.argv.join('='))) {
+        return;
       }
-    },
 
-    AssignmentExpression: function(node) {
-      if (isEnabled('assignment')) {
-        check(node.right);
-      }
-    },
-
-    ReturnStatement: function(node) {
-      if (isEnabled('return')) {
-        check(node.argument);
-      }
+      /* eslint-disable no-console */
+      console.log('The react/wrap-multilines rule is deprecated. Please ' +
+                  'use the react/jsx-wrap-multilines rule instead.');
+      /* eslint-enable no-console */
+      isWarnedForDeprecation = true;
     }
-  };
-
+  });
 };
 
 module.exports.schema = [{
